Add GitIgnore, Use Lookup Table, Now It's 4OP!
The biggest change is that now this is a 4OP synth! In order to do
that without getting output underflows, though, we needed to use a
sine lookup table rather than a sine trig function. So that change
was made, and the range of modulation indexes was raised to 500 in
order to get more interesting sounds out of it. Also added the file
which generated the sine lookup table, and another file to test it
by generating a sine wav file, and a .gitignore which ignores any
generated wav files we might generate as the project moves forward.