Revert Oppo And Make Exponential Envelope In Sine Wav Demo
Reverts Oppo to the last functional version, and uses the exponential
table to generate an envelope in the sine wave demo. This is still
using multiplications and divisions, though, so we need to figure
out how to set an accumulator or something instead to avoid all those
cycles being spent calculating the current amplitude values for the
envelope. We have chased the multiplications out of the main sine
equation and back into the envelope generator, and once we chase them
out of there we will have a working NCO.